requirements.txt
faiss-cpu>=1.8.0,<2
numpy>=1.26,<3
requests>=2.31,<3
sentence-transformers>=2.7.0,<4
